fixed time overflow

+++ devices/src/physical/ntp.h
@@ -3,6 +3,6 @@
 
 const char* ntpServer = "pool.ntp.org";
 
-unsigned long getTime();
+unsigned long long getTime();
 
 #endif
